Split index.js
