+++ src/defaultToCursor.js
@@ -1,0 +1,9 @@
+import base64 from 'base-64'
+import utf8 from 'utf8'
+
+function mrDefaultToCursor (field, id) {
+  const m = { field, id }
+  return base64.encode(utf8.encode(JSON.stringify(m)))
+}
+
+export default mrDefaultToCursor
