+++ src/leaf.js
@@ -1,0 +1,5 @@
+function leaf (object, path) {
+  return path.split('.').reduce((value, key) => value[key], object)
+}
+
+export default leaf
