+++ src/type.js
@@ -1,0 +1,23 @@
+import { connectionDefinitions } from 'graphql-relay'
+import { GraphQLInt } from 'graphql'
+
+/**
+ * Create a type called `${name}Connection`.
+ * @param {string} name Prefix name of connection.
+ * @param {object} nodeType GraphQLObjectType of the underlying node.
+ */
+function mrType (name, nodeType) {
+  const { connectionType } = connectionDefinitions({
+    name,
+    nodeType,
+    connectionFields: {
+      totalCount: {
+        type: GraphQLInt,
+        resolve: conn => conn.totalCount
+      }
+    }
+  })
+  return connectionType
+}
+
+export default mrType
